+++ Scripts/web_image.gd
@@ -1,16 +1,16 @@
 class_name WebImage
 
 static func load_image_from_url(url: String, target:TextureRect):
-	Logger.info("Loading image from URL: " + str(url))
+	AppLogger.info("Loading image from URL: " + str(url))
 	var http_request = HTTPRequest.new()
 	target.add_child(http_request)
 	http_request.request_completed.connect(on_request_completed.bind(target))
 	var err = http_request.request(url)
 	if err != OK:
-		Logger.error("Failed to load image from URL: " + url)
+		AppLogger.error("Failed to load image from URL: " + url)
 		return
 	else:
-		# Logger.info("Requested image...")
+		# AppLogger.info("Requested image...")
 		pass
 	
 static func on_request_completed(result: int, _response_code: int, headers: PackedStringArray, body:PackedByteArray, target:TextureRect):
@@ -27,9 +27,9 @@ static func on_request_completed(result: int, _response_code: int, headers: Pack
 			"image/gif":
 				load_as_gif(body, target)
 			_:
-				Logger.error("Unknown mime type: " + str(mime_type))
+				AppLogger.error("Unknown mime type: " + str(mime_type))
 	else:
-		Logger.error("Failed to load image from URL: " + str(result))
+		AppLogger.error("Failed to load image from URL: " + str(result))
 
 static func get_mime_type(headers: PackedStringArray) -> String:
 	for header in headers:
@@ -39,21 +39,21 @@ static func get_mime_type(headers: PackedStringArray) -> String:
 	return "unknown"
 	
 static func load_as_png(buffer:PackedByteArray, target:TextureRect):
-	Logger.info("Loading texture as PNG...")
+	AppLogger.info("Loading texture as PNG...")
 	var image = Image.new()
 	var err = image.load_png_from_buffer(buffer)
 	if err != OK:
-		Logger.error("Failed to parse PNG image from data: " + str(err))
+		AppLogger.error("Failed to parse PNG image from data: " + str(err))
 		return
 	
 	target.texture = ImageTexture.create_from_image(image)
 
 static func load_as_jpg(buffer:PackedByteArray, target:TextureRect):
-	Logger.info("Loading texture as JPG...")
+	AppLogger.info("Loading texture as JPG...")
 	var image = Image.new()
 	var err = image.load_jpg_from_buffer(buffer)
 	if err != OK:
-		Logger.error("Failed to parse JPEG image from data: " + str(err))
+		AppLogger.error("Failed to parse JPEG image from data: " + str(err))
 		return
 	
 	target.texture = ImageTexture.create_from_image(image)
@@ -72,7 +72,7 @@ static func load_from_path(path:String, target:TextureRect):
 		var image = Image.new()
 		var err = image.load(path)
 		if err != OK:
-			Logger.error("Failed to parse image from path: " + path)
+			AppLogger.error("Failed to parse image from path: " + path)
 			return
 		target.texture = ImageTexture.create_from_image(image)